pyproject.toml
[tool.poetry]
name = "repl-nix-dlthealthadvisor"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.39.0"
plotly = "^5.24.1"
psycopg2-binary = "^2.9.9"
matplotlib = "^3.9.2"
pandas = "^2.2.3"
scikit-learn = "^1.5.2"
flask-login = "^0.6.3"
werkzeug = "^3.0.4"
flask = "^3.0.3"
requests = "^2.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
